Finished assignment six for layout css
Looked over solution and I got everything except
for the footer. The footer was not going to the
bottom of the screen and staying there while
I resized my window. After looking at the solution
they actually applied a column layout to the
entire <body> of the webpage, then they set a
flex: 1; property to the .content to fill the
page, pushing the footer to the bottom because
the content in the middle (the odin logo and
search bar and buttons fill the entire page)