Finished assignment seven with solution notes
Put solution notes in css but successfully copied
the "holy grail" layout with only some minor
misses when compared to solution. Good stuff.
Remember, lists have initial padding and margin
you need to account for and content is a group
and items are individual items. Align means
position along the cross axis and the
justify means to position along the main axis.